修复拼写错误；属性flaot用CGflaot替换

+++ Tap Run/MKPlatform.h
@@ -10,7 +10,7 @@
 @class MKBlock;
 @interface MKPlatform : SKNode
 @property(nonatomic, strong) NSMutableArray* blocks;
-@property(nonatomic, assign) float speed;
+@property(nonatomic, assign) CGFloat speed;
 @property(nonatomic, readonly, assign) uint32_t contactTestBitMask;
 @property(nonatomic, readonly, assign) uint32_t categoryBitMask;
 /**
